+++ handle_other_flags.c
@@ -6,7 +6,7 @@
 /*   By: bkandemi <bkandemi@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/22 13:39:41 by bkandemi          #+#    #+#             */
-/*   Updated: 2022/03/24 09:05:01 by bkandemi         ###   ########.fr       */
+/*   Updated: 2022/03/25 09:05:38 by bkandemi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	handle_flags(t_flag *flag, char *str, int negative)
 			return (write(1, "0x", 2));
 		if (flag->spec == 'X')
 			return (write(1, "0X", 2));
-		if (flag->spec == 'o' && flag->prec <= ft_strlen(str))
+		if (flag->spec == 'o' && flag->prec <= (int)ft_strlen(str))
 			return (write(1, "0", 1));
 	}
 	return (0);
@@ -37,7 +37,7 @@ int	handle_flags(t_flag *flag, char *str, int negative)
 int	handle_plus_space_hash(t_flag *flag, char *str, int negative)
 {
 	int			ret;
-	uintmax_t	len;
+	int	len;
 
 	ret = handle_flags(flag, str, negative);
 	len = ft_strlen(str);
